requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pymongo==4.11.3
pydantic==2.10.6
pydantic-core==2.27.2
elasticsearch==8.17.1
jieba==0.42.1
anthropic==0.32.0
httpx==0.24.1
openai==1.3.0
typing-extensions==4.12.2
python-dotenv==1.0.1
coloredlogs==15.0.1
numpy==2.1.3
pandas==2.2.3
oss2 == 2.18.6
openpyxl == 3.1.5
zhipuai
motor == 3.7.0
xmltodict == 0.14.2
tabulate == 0.9.0
statsmodels==0.14.4
aiohttp == 3.11.16
anyio == 3.7.1
python-docx ==1.1.2
beautifulsoup4==4.12.3
markdownify==1.1.0
python-multipart==0.0.18
pypinyin==0.54.0
PyPDF2==3.0.1
